pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Maxim Bugaev <abc-java@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
pydantic-settings = "^2.7.1"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
psycopg2-binary = "^2.9.10"
fastapi-pagination = "^0.12.34"
aiohttp = "^3.11.11"

[tool.poetry.group.dev.dependencies]
anyio = "^3.6.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
faker = "^33.1.0"
pytest-cov = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
